chore(@clayui/modal): move content from index.tsx to Modal.tsx
